Fix script and bump kernel version to 6.4.8
The script had several issues which didn't allow it to run successfully.
Here's a summary of the changes:
- At least on Fedora, curl -O causes the script to output the tarball to
the current directory, not the location indicated, so I've removed that
flag
- The directory set for patch is invalid. Add a variable which gives an
absolute location
- patch will fail if it can't find a directory for a variant that isn't
stoney. Ignore the result and continue
- On stoney, zstd will detect file conflicts and hang when it comes to
any of them. Detect the file extensions and run the respective
decompression command for each archive type